refactor(backend): split guild service into three focused services
GuildService.ts mixed bot-presence caching, channel/emoji fetching, and
member/role context behind one 663-line class. Extracted the bot-status
domain (cache, invite URL, enrichment, guild details) into
GuildBotStatusService and the channel/emoji fetching into
GuildChannelService; GuildService now delegates and drops to ~215 lines
focused on member context and thin role delegates. Also moved the
duplicated GuildChannelOption/GuildEmojiOption types out of the backend
and frontend into packages/shared/src/types/guild.ts so both sides
import one definition.

Public API of guildService is unchanged (pure delegation), so no
consumer file needed edits. Verified against the existing 84
GuildService/guilds-route tests plus 92 tests across RoleGroupService,
GuildAccessService, rbac, admin, and roles routes, all passing
unmodified, and the full backend (87 suites/1410 tests) and frontend
(91 files/1051 tests) suites.

Mutation-tested the moved SSRF snowflake guard in getGuildEmojis and
found it has no existing test coverage (pre-existing gap, not a
regression from this move) - filed as #2269 rather than fixed inline
to keep this PR a pure, behavior-preserving refactor.
