+++ packages/backend/src/services/GuildService.ts
@@ -1,42 +1,34 @@
-import { z } from 'zod'
 import type { Client } from 'discord.js'
-import { BOT_INVITE_PERMISSIONS } from '@lucky/shared/constants'
 import { discordOAuthService, type DiscordGuild } from './DiscordOAuthService'
 import {
     setClient as setDiscordClient,
     getClient as getDiscordClient,
-    getServableGuild,
 } from '../utils/discordClientAccessor'
-import { metricsService } from './MetricsCache'
 import {
     roleService,
     type GuildRoleOption,
     type GuildRoleManage,
     type RoleUpsertData,
 } from './RoleService'
+import {
+    guildBotStatusService,
+    type GuildWithBotStatus,
+} from './GuildBotStatusService'
+import { guildChannelService } from './GuildChannelService'
 import { debugLog, errorLog } from '@lucky/shared/utils'
+import type { GuildChannelOption, GuildEmojiOption } from '@lucky/shared/types'
 import { isSnowflakeId } from '../schemas/common'
 
 const DISCORD_API_BASE_URL = 'https://discord.com/api/v10'
-const BOT_GUILD_CACHE_TTL_MS = 60_000
-// Discord permission bitfield literal representing zero permissions granted
-const NO_PERMISSIONS = '0'
 
 export interface GuildMemberContext {
     nickname: string | null
     roleIds: string[]
 }
 
-export interface GuildChannelOption {
-    id: string
-    name: string
-}
-
-export interface GuildEmojiOption {
-    id: string
-    name: string
-    animated: boolean
-}
+// Re-exported for backward compatibility with existing consumers.
+export type { GuildChannelOption, GuildEmojiOption } from '@lucky/shared/types'
+export type { GuildWithBotStatus } from './GuildBotStatusService'
 
 // Re-export role interfaces from RoleService for backward compatibility
 export type {
@@ -47,35 +39,12 @@ export type {
 
 export function setBotClient(client: Client | null): void {
     setDiscordClient(client)
-    guildService.clearBotGuildCache()
-}
-
-export interface GuildWithBotStatus extends DiscordGuild {
-    hasBot: boolean
-    botInviteUrl?: string
-    memberCount: number | null
-    categoryCount: number | null
-    textChannelCount: number | null
-    voiceChannelCount: number | null
-    roleCount: number | null
+    guildBotStatusService.clearBotGuildCache()
 }
 
 class GuildService {
-    private botGuildIdsCache: {
-        guildIds: Set<string>
-        expiresAt: number
-    } | null = null
-
-    private botGuildIdsInFlight: Promise<Set<string> | null> | null = null
-
     private getBotClient(): Client | null {
         return getDiscordClient()
-    }
-
-    clearBotGuildCache(): void {
-        this.botGuildIdsCache = null
-        this.botGuildIdsInFlight = null
-        metricsService.clearCache()
     }
 
     private getBotToken(): string | null {
@@ -83,139 +52,8 @@ class GuildService {
         return token && token.length > 0 ? token : null
     }
 
-    private validateChannelArray(data: unknown): Array<{
-        id?: string
-        name?: string
-        type: number
-        position?: number
-    }> {
-        if (!Array.isArray(data)) {
-            errorLog({
-                message: 'Invalid channels response from Discord API',
-                data: { expectedArray: true, receivedType: typeof data },
-            })
-            return []
-        }
-
-        const channelSchema = z.object({
-            id: z.string().optional(),
-            name: z.string().optional(),
-            type: z.number().int(),
-            position: z.number().optional(),
-        })
-
-        const validated: Array<{
-            id?: string
-            name?: string
-            type: number
-            position?: number
-        }> = []
-        for (const item of data) {
-            const result = channelSchema.safeParse(item)
-            if (result.success) {
-                validated.push(result.data)
-            } else {
-                debugLog({
-                    message: 'Skipping invalid channel in Discord API response',
-                    data: { errors: result.error.issues },
-                })
-            }
-        }
-        return validated
-    }
-
-    private async fetchBotGuildIds(token: string): Promise<Set<string> | null> {
-        try {
-            const response = await fetch(
-                `${DISCORD_API_BASE_URL}/users/@me/guilds`,
-                {
-                    headers: {
-                        Authorization: `Bot ${token}`,
-                    },
-                    signal: AbortSignal.timeout(10_000),
-                },
-            )
-
-            if (!response.ok) {
-                const responseBody = await response.text()
-                errorLog({
-                    message: 'Failed to fetch bot guilds from Discord API',
-                    data: {
-                        status: response.status,
-                        responseBody,
-                    },
-                })
-                return null
-            }
-
-            const payload = (await response.json()) as unknown
-            if (!Array.isArray(payload)) {
-                errorLog({
-                    message: 'Invalid bot guild payload from Discord API',
-                })
-                return null
-            }
-
-            const guildIds = new Set<string>()
-            for (const item of payload) {
-                if (
-                    typeof item === 'object' &&
-                    item !== null &&
-                    typeof (item as { id?: unknown }).id === 'string'
-                ) {
-                    guildIds.add((item as { id: string }).id)
-                }
-            }
-
-            this.botGuildIdsCache = {
-                guildIds,
-                expiresAt: Date.now() + BOT_GUILD_CACHE_TTL_MS,
-            }
-
-            debugLog({
-                message: 'Fetched bot guild ids from Discord API',
-                data: { guildCount: guildIds.size },
-            })
-
-            return guildIds
-        } catch (error) {
-            errorLog({
-                message: 'Error fetching bot guild ids from Discord API',
-                error,
-            })
-            return null
-        }
-    }
-
-    private async getBotGuildIds(): Promise<Set<string> | null> {
-        const token = this.getBotToken()
-        if (!token) {
-            this.clearBotGuildCache()
-            return null
-        }
-
-        const now = Date.now()
-        if (this.botGuildIdsCache && this.botGuildIdsCache.expiresAt > now) {
-            return this.botGuildIdsCache.guildIds
-        }
-
-        if (this.botGuildIdsInFlight) {
-            return this.botGuildIdsInFlight
-        }
-
-        this.botGuildIdsInFlight = this.fetchBotGuildIds(token).finally(() => {
-            this.botGuildIdsInFlight = null
-        })
-
-        return this.botGuildIdsInFlight
-    }
-
     async hasBotInGuild(guildId: string): Promise<boolean> {
-        const botGuildIds = await this.getBotGuildIds()
-        return (
-            this.checkBotInGuild(guildId) ||
-            (botGuildIds?.has(guildId) ?? false)
-        )
+        return guildBotStatusService.hasBotInGuild(guildId)
     }
 
     async getGuildMemberContext(
@@ -296,191 +134,11 @@ class GuildService {
     async getGuildTextChannelOptions(
         guildId: string,
     ): Promise<GuildChannelOption[]> {
-        if (!isSnowflakeId(guildId)) {
-            return []
-        }
-
-        const client = this.getBotClient()
-
-        if (client) {
-            try {
-                const guild =
-                    client.guilds.cache.get(guildId) ??
-                    (await client.guilds.fetch(guildId))
-                const channels = await guild.channels.fetch()
-                return [...channels.values()]
-                    .filter(
-                        (channel): channel is NonNullable<typeof channel> =>
-                            channel !== null &&
-                            (channel.type === 0 ||
-                                channel.type === 5 ||
-                                channel.type === 15 ||
-                                channel.type === 16),
-                    )
-                    .sort((a, b) => a.rawPosition - b.rawPosition)
-                    .map((channel) => ({
-                        id: channel.id,
-                        name: `#${channel.name}`,
-                    }))
-            } catch (error) {
-                debugLog({
-                    message: 'Failed to fetch guild channels from bot client',
-                    error,
-                })
-            }
-        }
-
-        const token = this.getBotToken()
-        if (!token) {
-            return []
-        }
-
-        try {
-            const response = await fetch(
-                `${DISCORD_API_BASE_URL}/guilds/${guildId}/channels`,
-                {
-                    headers: {
-                        Authorization: `Bot ${token}`,
-                    },
-                    signal: AbortSignal.timeout(10_000),
-                },
-            )
-
-            if (!response.ok) {
-                return []
-            }
-
-            const payload = this.validateChannelArray(await response.json())
-
-            const filteredChannels = payload.filter(
-                (
-                    channel,
-                ): channel is {
-                    id: string
-                    name: string
-                    type: number
-                    position?: number
-                } =>
-                    typeof channel.id === 'string' &&
-                    typeof channel.name === 'string' &&
-                    (channel.type === 0 ||
-                        channel.type === 5 ||
-                        channel.type === 15 ||
-                        channel.type === 16),
-            )
-
-            return filteredChannels
-                .sort(
-                    (
-                        a: {
-                            id: string
-                            name: string
-                            type: number
-                            position?: number
-                        },
-                        b: {
-                            id: string
-                            name: string
-                            type: number
-                            position?: number
-                        },
-                    ) => (a.position ?? 0) - (b.position ?? 0),
-                )
-                .map(
-                    (channel: {
-                        id: string
-                        name: string
-                        type: number
-                        position?: number
-                    }) => ({
-                        id: channel.id,
-                        name: `#${channel.name}`,
-                    }),
-                )
-        } catch (error) {
-            errorLog({
-                message: 'Failed to fetch guild channels',
-                error,
-            })
-            return []
-        }
+        return guildChannelService.getGuildTextChannelOptions(guildId)
     }
 
     async getGuildEmojis(guildId: string): Promise<GuildEmojiOption[]> {
-        const client = this.getBotClient()
-
-        if (client) {
-            try {
-                const guild = await getServableGuild(guildId)
-                if (guild) {
-                    return [...guild.emojis.cache.values()].map((emoji) => ({
-                        id: emoji.id,
-                        name: emoji.name ?? '',
-                        animated: emoji.animated ?? false,
-                    }))
-                }
-            } catch (error) {
-                debugLog({
-                    message: 'Failed to fetch guild emojis from bot client',
-                    error,
-                })
-            }
-        }
-
-        const token = this.getBotToken()
-        if (!token) {
-            return []
-        }
-
-        // Reject anything that is not a Discord snowflake before it reaches the
-        // request URL — validated inline at the sink so the ID cannot forge the
-        // request (SSRF / path-traversal guard).
-        if (!/^\d{17,20}$/.test(guildId)) {
-            throw new Error('Invalid Discord guild id')
-        }
-
-        try {
-            const response = await fetch(
-                `${DISCORD_API_BASE_URL}/guilds/${guildId}/emojis`,
-                {
-                    headers: {
-                        Authorization: `Bot ${token}`,
-                    },
-                    signal: AbortSignal.timeout(10_000),
-                },
-            )
-
-            if (!response.ok) {
-                throw new Error(`Discord API error: ${response.status}`)
-            }
-
-            const payload = (await response.json()) as unknown[]
-
-            return payload
-                .filter(
-                    (
-                        emoji,
-                    ): emoji is {
-                        id: string
-                        name?: string
-                        animated?: boolean
-                    } =>
-                        typeof emoji === 'object' &&
-                        emoji !== null &&
-                        typeof (emoji as { id?: unknown }).id === 'string',
-                )
-                .map((emoji) => ({
-                    id: emoji.id,
-                    name: emoji.name ?? '',
-                    animated: emoji.animated ?? false,
-                }))
-        } catch (error) {
-            errorLog({
-                message: 'Failed to fetch guild emojis',
-                error,
-            })
-            throw error
-        }
+        return guildChannelService.getGuildEmojis(guildId)
     }
 
     async getUserGuilds(accessToken: string): Promise<DiscordGuild[]> {
@@ -500,102 +158,21 @@ class GuildService {
     }
 
     checkBotInGuild(guildId: string): boolean {
-        const client = this.getBotClient()
-        if (!client) {
-            return false
-        }
-
-        const guild = client.guilds.cache.get(guildId)
-        return guild !== undefined
+        return guildBotStatusService.checkBotInGuild(guildId)
     }
 
     generateBotInviteUrl(guildId?: string): string {
-        const clientId = process.env.CLIENT_ID
-        if (!clientId) {
-            throw new Error('CLIENT_ID not configured')
-        }
-
-        const scopes = ['bot', 'applications.commands']
-        // Was a hardcoded '8' — Administrator. The dashboard's "add Lucky to
-        // this server" flow asked every owner for full admin, contradicting
-        // both the ADR and the public listings (#1923). Shares the curated set
-        // with the landing page and the /invite redirect.
-        const permissions = BOT_INVITE_PERMISSIONS
-        const redirectUri = process.env.WEBAPP_REDIRECT_URI
-
-        let inviteUrl = `https://discord.com/api/oauth2/authorize?client_id=${clientId}&permissions=${permissions}&scope=${scopes.join('%20')}`
-
-        if (guildId) {
-            inviteUrl += `&guild_id=${guildId}`
-        }
-
-        if (redirectUri) {
-            inviteUrl += `&redirect_uri=${encodeURIComponent(redirectUri)}`
-        }
-
-        return inviteUrl
+        return guildBotStatusService.generateBotInviteUrl(guildId)
     }
 
     async enrichGuildsWithBotStatus(
         guilds: DiscordGuild[],
     ): Promise<GuildWithBotStatus[]> {
-        const botGuildIds = await this.getBotGuildIds()
-        const enrichedGuilds = await Promise.all(
-            guilds.map(async (guild) => {
-                const hasBot =
-                    this.checkBotInGuild(guild.id) ||
-                    (botGuildIds?.has(guild.id) ?? false)
-                const botInviteUrl = hasBot
-                    ? undefined
-                    : this.generateBotInviteUrl(guild.id)
-                const metrics = hasBot
-                    ? await metricsService.getGuildMetrics(guild.id)
-                    : {
-                          memberCount: null,
-                          categoryCount: null,
-                          textChannelCount: null,
-                          voiceChannelCount: null,
-                          roleCount: null,
-                      }
-
-                return {
-                    ...guild,
-                    hasBot,
-                    botInviteUrl,
-                    ...metrics,
-                }
-            }),
-        )
-
-        return enrichedGuilds
+        return guildBotStatusService.enrichGuildsWithBotStatus(guilds)
     }
 
     async getGuildDetails(guildId: string): Promise<GuildWithBotStatus | null> {
-        const client = this.getBotClient()
-        if (!client) {
-            return null
-        }
-
-        const guild = client.guilds.cache.get(guildId)
-        if (!guild) {
-            return null
-        }
-
-        const hasBot = true
-        const botInviteUrl = this.generateBotInviteUrl(guildId)
-        const metrics = await metricsService.getGuildMetrics(guildId)
-
-        return {
-            id: guild.id,
-            name: guild.name,
-            icon: guild.icon,
-            owner: false,
-            permissions: NO_PERMISSIONS,
-            features: guild.features,
-            hasBot,
-            botInviteUrl,
-            ...metrics,
-        }
+        return guildBotStatusService.getGuildDetails(guildId)
     }
 
     async getAllBotGuilds(): Promise<
@@ -609,31 +186,7 @@ class GuildService {
             roleCount: number | null
         }>
     > {
-        const client = this.getBotClient()
-        if (!client) {
-            return []
-        }
-
-        const guilds = [...client.guilds.cache.values()]
-
-        return Promise.all(
-            guilds.map(async (guild) => {
-                const metrics = await metricsService.getGuildMetrics(guild.id)
-                const iconUrl = guild.icon
-                    ? `https://cdn.discordapp.com/icons/${guild.id}/${guild.icon}.png?size=64`
-                    : null
-
-                return {
-                    id: guild.id,
-                    name: guild.name,
-                    iconUrl,
-                    memberCount: metrics.memberCount,
-                    textChannelCount: metrics.textChannelCount,
-                    voiceChannelCount: metrics.voiceChannelCount,
-                    roleCount: metrics.roleCount,
-                }
-            }),
-        )
+        return guildBotStatusService.getAllBotGuilds()
     }
 
     async getFullGuildRoles(guildId: string): Promise<GuildRoleManage[]> {
