+++ packages/shared/src/types/guild.ts
@@ -1,0 +1,10 @@
+export interface GuildChannelOption {
+    id: string
+    name: string
+}
+
+export interface GuildEmojiOption {
+    id: string
+    name: string
+    animated: boolean
+}
