+++ packages/shared/src/types/index.ts
@@ -2,4 +2,5 @@ export * from './featureToggle'
 export * from './errors'
 export * from './common/BaseResult'
 export * from './music'
+export * from './guild'
 export { RecommendationSource } from '../generated/prisma/enums'
